Config i3bar properties
The following i3bar properties can now be configured:

color, min_width, align, separator, separator_block_width

If given in the `general` section, this will be the default for all
blocks.

If given in the block level config, it will override the `general`
section for the block.

If given in `show_block` dbus method, it will override block level
config and general config for the block.